✨add : 올림픽 rtk 유효성 검사 추가

+++ src/components/table/Table.jsx
@@ -24,7 +24,9 @@ function Table() {
   });
 
   const onDelete = (id) => {
-    dispatch(deleteMedal(id));
+    if (window.confirm('정말 삭제할까요?')) {
+      dispatch(deleteMedal(id));
+    }
   };
   return (
     <>
